Cree Utilisateur : Refactor - tout en anglais sauf pour les paths
