+++ src/main/java/claurendeau/hackqc/algo/Backend/controllers/UserController.java
@@ -1,4 +1,4 @@
-package claurendeau.hackqc.algo.Backend.controleurs;
+package claurendeau.hackqc.algo.Backend.controllers;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -7,9 +7,9 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping(path = "/utilisateur")
-public class UtilisateurControleur {
+public class UserController {
     @PostMapping("/cree")
-    public ResponseEntity<String> creeUtilisateur() {
+    public ResponseEntity<String> createUser() {
         return null;
     }
 }
